Fixed types for TimeAndDate struct.

+++ common_types.h
@@ -28,13 +28,13 @@
 
 typedef struct TimeAndDate 
 {
-	int year;
-	int month;
-	int day;
-	int dow;
-	int hour;
-	int min;
-	int sec;
+	uint16_t year;
+	uint8_t month;
+	uint8_t day;
+	uint8_t dow;
+	uint8_t hour;
+	uint8_t min;
+	uint8_t sec;
 } TimeAndDate;
 
 #endif
